+++ app/scripts/components/index.ts
@@ -2,3 +2,4 @@ import "./hierarchy-path-canvas.component";
 import "./level.component";
 import "./bundlelist-sidebar.component";
 import "./emuhierarchy.component";
+import "./settings.component";