+++ app/scripts/controllers/index.ts
@@ -5,6 +5,5 @@ import "./manual.controller";
 import "./modal.controller";
 import "./search-annot.controller";
 import "./show-hierarchy.controller";
-import "./spect-settings.controller";
 import "./tabbed.controller";
 import "./tabbed-help.controller";