+++ webpack.common.js
@@ -48,6 +48,18 @@ module.exports = {
                         // Compiles Sass to CSS
                         "sass-loader"
                     ]
+                },
+                {
+                    test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
+                    use: [
+                      {
+                        loader: 'file-loader',
+                        options: {
+                          name: '[name].[ext]',
+                          outputPath: 'fonts/'
+                        }
+                      }
+                    ]
                 }
             ]
         }